🐞 Fix comet HPO (#597)
* Fix comet hpo + refactoring + fix metriccallback in benchmarking

* Move sweep runners + utils to anomalib
