+++ anomalib/utils/hpo/__init__.py
@@ -3,6 +3,6 @@
 # Copyright (C) 2022 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
 
-from .config import flatten_hpo_params
+from .runners import CometSweep, WandbSweep
 
-__all__ = ["flatten_hpo_params"]
+__all__ = ["CometSweep", "WandbSweep"]
